1.增加udp服务超时判断处理逻辑;\n2.增加bind异常时的逻辑判断处理,增加打印日志方便定位问题;

+++ PacketParse_Interface/PacketParse_Interface.cpp
@@ -55,8 +55,8 @@ bool parse_packet_from_recv_buffer_stream(uint32 connect_id, CSessionBuffer* buf
         }
 
         //如果有完整的包头
-        uint32 packet_version = 0;             //协议版本号
-        uint16 command_id = 0;             //CommandID
+        uint16 packet_version = 0;                 //协议版本号
+        uint16 command_id = 0;                     //CommandID
         uint32 packet_body_length = 0;             //包体长度
         char   packet_session[33] = { '\0' };      //Session字符串
 
@@ -137,7 +137,7 @@ bool parse_packet_from_recv_buffer_single(uint32 connect_id, CSessionBuffer* buf
     }
 
     //如果有完整的包头
-    uint32 packet_version = 0;             //协议版本号
+    uint16 packet_version = 0;             //协议版本号
     uint16 command_id = 0;             //CommandID
     uint32 packet_body_length = 0;             //包体长度
     char   packet_session[33] = { '\0' };      //Session字符串
